# lib/api-zod/src/generated/types/createProductRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Ethnic Clothing Store API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateProductRequest {
  name: string;
  description?: string;
  price: number;
  originalPrice?: number;
  categoryId: number;
  images?: string[];
  stock: number;
  sizes?: string[];
  colors?: string[];
  isFeatured?: boolean;
  isActive?: boolean;
}