# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Ethnic Clothing Store API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from "./orderItem";
import type { OrderPaymentMethod } from "./orderPaymentMethod";
import type { OrderPaymentStatus } from "./orderPaymentStatus";
import type { OrderStatus } from "./orderStatus";
import type { ShippingAddress } from "./shippingAddress";

export interface Order {
  id: number;
  userId: number;
  userName?: string;
  items: OrderItem[];
  subtotal: number;
  codCharge?: number;
  total: number;
  paymentMethod: OrderPaymentMethod;
  paymentStatus?: OrderPaymentStatus;
  status: OrderStatus;
  shippingAddress: ShippingAddress;
  createdAt: Date;
}